+++ core/src/main/java/com/ensody/reactivestate/CleanupCore.kt
@@ -4,11 +4,36 @@ import kotlinx.coroutines.*
 import kotlin.coroutines.CoroutineContext
 
 /**
- * Interface for an object that can be disposed/deactivated.
+ * An object that can be disposed/deactivated/canceled by calling `dispose()`.
  *
  * This is an alias for [DisposableHandle].
  */
 typealias Disposable = DisposableHandle
+
+/**
+ * A [Disposable] that can have additional `Disposable`s attached to it, so they are automatically
+ * disposed together with this object.
+ */
+interface AttachedDisposables : Disposable {
+    /** The attached disposables which should be auto-disposed when this object is disposed. */
+    val attachedDisposables: DisposableGroup
+}
+
+/**
+ * A [Disposable] executing the given [function] on `dispose()`.
+ *
+ * Example:
+ *
+ * ```kotlin
+ * val disposable = OnDispose { println("disposing myself") }
+ * disposable.dispose() // => "disposing myself"
+ * ```
+ */
+class OnDispose(val function: () -> Unit) : Disposable {
+    override fun dispose() {
+        function()
+    }
+}
 
 /** A [Disposable] wrapping a [Job]. */
 class JobDisposable(val job: Job) : Disposable {