+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * Added `thisWorkQueue` helper for passing an arg via `this`.
 * `argWorkQueue` now supports suspension functions.
+* Added `AttachedDisposables` interface for objects that can clean up other disposables.
+* Added `OnDispose { ... }` class for triggering a function when its `dispose()` method is called.
 * Further documentation improvements.
 
 ## 0.9.1